requirements.txt
aiofiles==24.1.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
Authlib==1.6.0
bandit==1.8.6
bcrypt==4.3.0
black==25.1.0
certifi==2025.7.9
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
click==8.2.1
coverage==7.9.2
cryptography==45.0.5
databases==0.9.0
distlib==0.3.9
dnspython==2.7.0
dparse==0.6.4
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
filelock==3.16.1
flake8==7.3.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.12
idna==3.10
iniconfig==2.1.0
isort==6.0.1
Jinja2==3.1.6
joblib==1.5.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
mccabe==0.7.0
mdurl==0.1.2
mypy==1.17.0
mypy_extensions==1.1.0
nltk==3.9.1
nodeenv==1.9.1
packaging==25.0
passlib==1.7.4
pathspec==0.12.1
pbr==6.1.1
platformdirs==4.3.8
pluggy==1.6.0
pre_commit==4.2.0
psutil==6.1.1
pyasn1==0.6.1
pycodestyle==2.14.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.1.0
pytest-cov==6.2.1
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.8
rignore==0.6.2
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
ruff==0.12.3
safety==3.6.0
safety-schemas==0.0.14
sentry-sdk==2.33.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
stevedore==5.4.1
tenacity==9.1.2
tomlkit==0.13.3
tqdm==4.67.1
typer==0.16.0
types-passlib==1.7.7.20250602
types-pyasn1==0.6.0.20250516
types-python-jose==3.5.0.20250531
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.1.0
websockets==15.0.1
